renaming operator->operation for consistency
