+++ src/wmtk/utils/OperationRecordingDataTypes.cpp
@@ -1,4 +1,4 @@
-#include <wmtk/utils/OperatorRecordingDataTypes.hpp>
+#include <wmtk/utils/OperationRecordingDataTypes.hpp>
 
 using namespace wmtk;
